pyproject.toml
[project]
name = "clap-rap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discogs-client>=2.3.0",
    "jupyter>=1.1.1",
    "lyricsgenius>=3.7.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "wikipedia-api>=0.8.1",
    "yt-dlp>=2025.12.8",
]
